Refactor: common hapi/codec.py
