+++ speasy/core/codecs/bundled_codecs/hapi/__init__.py
@@ -1,1 +1,4 @@
 from .csv import HapiCsv
+
+
+
